move localstorage methods from utils.js to its own file
